requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
azure-core==1.36.0
azure-cosmos==4.14.2
blinker==1.9.0
cachelib==0.13.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
distro==1.9.0
Flask==3.1.2
frozenlist==1.8.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
Hypercorn==0.18.0
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
MarkupSafe==3.0.3
msal==1.34.0
msgspec==0.19.0
multidict==6.7.0
openai==2.8.1
packaging==25.0
priority==2.0.0
propcache==0.4.1
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
Quart==0.20.0
requests==2.32.5
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.3.2
yarl==1.22.0
